Stats upload unhappy invalid folder record

+++ StatsDownload/StatsDownload.Core/DataTransfer/UserData.cs
@@ -2,15 +2,11 @@
 {
     public class UserData
     {
-        /// <summary>
-        ///     The intent of this constructor is to be used during testing when one does not care about the values in the DTO
-        ///     This constructor does not modify the properties, leaving them the default value
-        /// </summary>
         public UserData()
         {
         }
 
-        public UserData(string name, ulong totalPoints, int totalWorkUnits, int teamNumber)
+        public UserData(string name, ulong totalPoints, ulong totalWorkUnits, ulong teamNumber)
         {
             Name = name;
             TotalPoints = totalPoints;
@@ -24,10 +20,10 @@
 
         public string Name { get; private set; }
 
-        public int TeamNumber { get; private set; }
+        public ulong TeamNumber { get; private set; }
 
         public ulong TotalPoints { get; private set; }
 
-        public int TotalWorkUnits { get; private set; }
+        public ulong TotalWorkUnits { get; private set; }
     }
 }
